backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
itsdangerous>=2.1.2

# Browser Automation
selenium>=4.15.0
playwright>=1.40.0
webdriver-manager>=4.0.1

# AI/LLM
openai>=1.6.0
google-genai>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0

# Audio/Video Processing
openai-whisper>=20231117
torch>=2.1.0
torchaudio>=2.1.0
pydub>=0.25.1
moviepy>=1.0.3
opencv-python>=4.8.1

# Document Processing
openpyxl>=3.1.2
pandas>=2.1.4
pdfplumber>=0.10.3
PyPDF2>=3.0.1

# Vector Search & Embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3

# Date & Time
python-dateutil>=2.8.2
holidays>=0.38

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0

# CLI
click>=8.1.7

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
cryptography>=41.0.7
requests>=2.31.0
websockets>=12.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-selenium>=4.0.1
pytest-cov>=4.1.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
